requirements.txt
pandas
google-cloud-bigquery
scikit-learn
matplotlib
seaborn
google-api-python-client
db-dtypes